hms-enterprise-grade/backend/requirements.txt
Django==4.2.23
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
django-filter==24.3
django-cors-headers==4.7.0
python-dotenv==1.1.1
psycopg2-binary==2.9.10
boto3==1.34.162
django-storages==1.14.4
celery==5.4.0
redis==5.0.8
django-redis==5.4.0
django-celery-beat==2.7.0
sentry-sdk==2.13.0
whitenoise==6.7.0
django-prometheus==2.3.1
argon2-cffi==23.1.0
django-cryptography==1.1
openpyxl==3.1.5
lxml==5.3.0
requests==2.32.3
pybreaker==1.0.2
cryptography==43.0.1
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-django
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-psycopg2
kafka-python==2.0.2
six==1.16.0
django-encrypted-model-fields==0.6.5
